+++ l5kit/l5kit/dataset/__init__.py
@@ -1,6 +1,6 @@
 from .ego import EgoDataset
 from .agent import AgentDataset
 
-from .utilities import build_dataloader
+from .dataloader_builder import build_dataloader
 
 __all__ = ["EgoDataset", "AgentDataset", "build_dataloader"]
