+++ l5kit/l5kit/rasterization/__init__.py
@@ -1,6 +1,6 @@
 from .box_rasterizer import BoxRasterizer
 from .combine_rasterizer import CombineRasterizer
-from .factory import build_rasterizer
+from .rasterizer_builder import build_rasterizer
 from .rasterizer import Rasterizer, EGO_EXTENT_WIDTH, EGO_EXTENT_LENGTH, EGO_EXTENT_HEIGHT
 from .sat_box_rasterizer import SatBoxRasterizer
 from .satellite_image import get_sat_image_crop, get_sat_image_crop_scaled, get_sat_image_crop_scaled_from_ecef
