+++ l5kit/l5kit/visualization/__init__.py
@@ -1,5 +1,5 @@
 from .gif import write_gif
-from .utilities import PREDICTED_POINTS_COLOR, TARGET_POINTS_COLOR, draw_reference_trajectory, draw_trajectory
+from .utils import PREDICTED_POINTS_COLOR, TARGET_POINTS_COLOR, draw_reference_trajectory, draw_trajectory
 from .video import write_video
 
 __all__ = [
